requirements.txt
langchain-google-genai
google-generativeai
prefect
boto3
localstack-client
fastapi
uvicorn
python-dotenv
langchain
python-dotenv